I made the following changes.
There are 6 flag:
1) Collection: react to studio tag
2) Studio: react to studio tag
3) Source: react to the file name
4) Resolution: react to studio tag
5) Audio: react to studio tag
6) Language: react to studio tag

To make it easier to handle all of the tag I made few changes to the aeon skin structure.
1) I created 6 new xml files
1.1) Collection: Includes_MediaFlags_collection.xml
1.2) Studio: Includes_MediaFlags_studio.xml
1.3) Source: Includes_MediaFlags_source.xml
1.4) Resolution: Includes_MediaFlags_video.xml
1.5) Audio: Includes_MediaFlags_audio.xml
1.6) Language: Includes_MediaFlags_language.xml
2) In the Includes_MediaFlags.xml then I use the reference from the include files above
3) Of course all these includes files are referred in the Includes.xml file
In this way it should be easier to handle the different icons and xml's per section without interfere with dhj or the other developer work.

Just to make an example, the "Film Studio Icons" from fernferret: now to use it you need to override the includes_mediaflag.xml (maybe losing a custom made change eg. to the audio section), with this mod you will have just to download the includes_mediaflag_studio.xml and overwrite it without worries (of course if fernferret made the little changes neesed), I think in this way it could easily becoming generated automaticaly on the fernferret web page.

Another modification I've done has been to organize all the images inside the flag directory so it looks like:
1) flags/audio: images for the audio flag
2) flags/collection: images for the collection flag
3) flags/info: untouched (for the info view)
4) flags/language: images for the language flag
5) flags/mpaa: images for the mpaa rating (needed to change the DialogVideoInfo.xml file)
6) flags/source: images for the source flag
7) flags/studio: images for the studio flag
8) flags/video: images for the resolution flag

I've started from the CF2009 mod so it has all his great mods inside (submenu, etc etc)...

Now few steps to improve it (of course if you guys agree with the approach):
1) CF2009, if you like you can integrate it in your overall mods
2) DHJ, for the final release maybe you could integrate this approach directly in aeon
3) fekker, to make it work moreasily it would be nice if you can add always the string "lan" before the languages in the studio tag (just like the "sub" string for the subtitles) these will help make easier to handle the language, subtitle tag
4) fernferret, if you'd like you could make your web page create a file like the includes_mediaflag_studio.xml